+++ src/Events/EventModel.php
@@ -9,13 +9,12 @@ use Carbon\CarbonImmutable;
 use Illuminate\Database\Eloquent\Model;
 
 /**
- * Class DurableEvent
- * @package Chocofamilyme\LaravelPubSub\Events
- *
  * @property string $id
  * @property string $name
  * @property array $payload
+ * @property array $headers
  * @property string $exchange
+ * @property string $exchange_type
  * @property string $routing_key
  * @property Carbon $created_at
  * @property Carbon|CarbonImmutable|string|null $processed_at
@@ -38,11 +37,41 @@ class EventModel extends Model
     ];
 
     protected $casts = [
-        'payload' => 'array'
+        'payload' => 'array',
+        'headers' => 'array',
     ];
+
+    protected ?PublishEvent $originalEvent = null;
 
     public function getTable()
     {
         return config('pubsub.tables.events', parent::getTable());
+    }
+
+    public function setOriginalEvent(PublishEvent $event): self
+    {
+        $this->originalEvent = $event;
+
+        return $this;
+    }
+
+    public function isDurable(): bool
+    {
+        if (null === $this->originalEvent) {
+            return false;
+        }
+
+        return $this->originalEvent instanceof DurableEvent;
+    }
+
+    public function amqpProperties(): array
+    {
+        return [
+            'exchange' => [
+                'name' => $this->exchange,
+                'type' => $this->exchange_type,
+            ],
+            'headers'  => $this->headers ?? [],
+        ];
     }
 }