+++ src/Providers/PubSubServiceProvider.php
@@ -1,26 +1,22 @@
 <?php
 
+declare(strict_types=1);
+
 namespace Chocofamilyme\LaravelPubSub\Providers;
 
-use Chocofamilyme\LaravelPubSub\Amqp\Amqp;
-use Chocofamilyme\LaravelPubSub\Amqp\AmqpFacade;
+use Chocofamilyme\LaravelPubSub\Commands\EventRePublish;
 use Chocofamilyme\LaravelPubSub\Commands\EventListenCommand;
-use Chocofamilyme\LaravelPubSub\Events\Dispatcher;
 use Chocofamilyme\LaravelPubSub\Listener;
 use Illuminate\Contracts\Debug\ExceptionHandler;
-use Illuminate\Filesystem\Filesystem;
-use Illuminate\Queue\QueueManager;
-use Illuminate\Contracts\Queue\Factory as QueueFactoryContract;
 use Illuminate\Support\Collection;
-use VladimirYuldashev\LaravelQueueRabbitMQ\LaravelQueueRabbitMQServiceProvider;
-use VladimirYuldashev\LaravelQueueRabbitMQ\Queue\Connectors\RabbitMQConnector;
+use Illuminate\Support\ServiceProvider;
 
 /**
  * Class PubSubServiceProvider
  *
  * @package Chocofamilyme\LaravelPubSub\Providers
  */
-class PubSubServiceProvider extends LaravelQueueRabbitMQServiceProvider
+class PubSubServiceProvider extends ServiceProvider
 {
     /**
      * Register services.
@@ -29,8 +25,6 @@ class PubSubServiceProvider extends LaravelQueueRabbitMQServiceProvider
      */
     public function register(): void
     {
-        parent::register();
-
         // Merge our config with application config
         $this->mergeConfigFrom(
             __DIR__ . '/../config/queue.php',
@@ -42,38 +36,41 @@ class PubSubServiceProvider extends LaravelQueueRabbitMQServiceProvider
         );
 
         if ($this->app->runningInConsole()) {
-            $this->app->singleton('rabbitmq.listener', function () {
-                $isDownForMaintenance = function (): bool {
-                    return $this->app->isDownForMaintenance();
-                };
+            $this->app->singleton(
+                'rabbitmq.listener',
+                function () {
+                    $isDownForMaintenance = function (): bool {
+                        return $this->app->isDownForMaintenance();
+                    };
 
-                /** @psalm-suppress UndefinedInterfaceMethod */
-                return new Listener(
-                    $this->app['queue'],
-                    $this->app['events'],
-                    $this->app[ExceptionHandler::class],
-                    $isDownForMaintenance
-                );
-            });
+                    /** @psalm-suppress UndefinedInterfaceMethod */
+                    return new Listener(
+                        $this->app['queue'],
+                        $this->app['events'],
+                        $this->app[ExceptionHandler::class],
+                        $isDownForMaintenance
+                    );
+                }
+            );
 
-            $this->app->singleton(EventListenCommand::class, static function ($app) {
-                return new EventListenCommand(
-                    $app['rabbitmq.listener'],
-                    $app['cache.store']
-                );
-            });
+            $this->app->singleton(
+                EventListenCommand::class,
+                static function ($app) {
+                    return new EventListenCommand(
+                        $app['rabbitmq.listener'],
+                        $app['cache.store']
+                    );
+                }
+            );
 
             // Register artisan commands
-            $this->commands([
-                EventListenCommand::class,
-            ]);
+            $this->commands(
+                [
+                    EventListenCommand::class,
+                    EventRePublish::class,
+                ]
+            );
         }
-
-        $this->app->singleton('events', function ($app) {
-            return (new Dispatcher($app))->setQueueResolver(function () use ($app) {
-                return $app->make(QueueFactoryContract::class);
-            });
-        });
     }
 
     /**
@@ -84,40 +81,29 @@ class PubSubServiceProvider extends LaravelQueueRabbitMQServiceProvider
     public function boot(): void
     {
         // Config
-        $this->publishes([
-            __DIR__ . '/../config/pubsub.php' => config_path('pubsub.php'),
-        ]);
-        $this->publishes([
-            __DIR__ . '/../../database/migrations/create_pubsub_events_table.php.stub' => $this->getMigrationFileName('_create_pubsub_events_table.php'),
-        ], 'migrations');
-        $this->publishes([
-            __DIR__ . '/../../database/migrations/update_pubsub_events_table_add_failed_columns.php.stub' => $this->getMigrationFileName('_update_pubsub_events_table_add_failed_columns.php'),
-        ], 'migrations');
-
-        // Add class and it's facade
-        $this->app->bind('Amqp', function ($app) {
-            return new Amqp($app['queue']);
-        });
-
-        if (!class_exists('Amqp')) {
-            class_alias(AmqpFacade::class, 'Amqp');
-        }
-
-        /**
-         * @var QueueManager $queue
-         * @psalm-suppress UndefinedInterfaceMethod
-         */
-        $queue = $this->app['queue'];
-
-        $queue->addConnector('rabbitmq', function () {
-            /** @psalm-suppress UndefinedInterfaceMethod */
-            return new RabbitMQConnector($this->app['events']);
-        });
-    }
-
-    public function provides()
-    {
-        return ['Amqp'];
+        $this->publishes(
+            [
+                __DIR__ . '/../config/pubsub.php' => config_path('pubsub.php'),
+            ]
+        );
+        $this->publishes(
+            [
+                __DIR__
+                . '/../../database/migrations/create_pubsub_events_table.php.stub' => $this->getMigrationFileName(
+                    '_create_pubsub_events_table.php'
+                ),
+            ],
+            'migrations'
+        );
+        $this->publishes(
+            [
+                __DIR__
+                . '/../../database/migrations/update_pubsub_events_table_add_failed_columns.php.stub' => $this->getMigrationFileName(
+                    '_update_pubsub_events_table_add_failed_columns.php'
+                ),
+            ],
+            'migrations'
+        );
     }
 
     /**