+++ tests/TestCase.php
@@ -1,13 +1,18 @@
 <?php
 
+declare(strict_types=1);
+
 namespace Chocofamily\LaravelPubSub\Tests;
 
+use Chocofamilyme\LaravelPubSub\Providers\PubSubServiceProvider;
 use Orchestra\Testbench\TestCase as Orchestra;
 
-/**
- * @package Chocolife.me
- * @author  Moldabayev Vadim <moldabayev.v@chocolife.kz>
- */
 abstract class TestCase extends Orchestra
 {
+    protected function setUp(): void
+    {
+        parent::setUp();
+
+        $this->app->register(PubSubServiceProvider::class);
+    }
 }